Fix importer & Split task CLI in two command lines Fix bug in UI
